# Realpath performance