--- a/tools/fs/optimistic.ts
+++ b/tools/fs/optimistic.ts
@@ -274,24 +274,6 @@ export const optimisticLStatOrNull = makeCheapPathFunction(
   },
 );
 
-export const optimisticRealpath = makeOptimistic('realpath', realpath);
-export const optimisticRealpathOrNull = makeOptimistic('realpathOrNull', (
-  path: string,
-  options?: Parameters<typeof optimisticRealpath>[1],
-) => {
-  try {
-    return realpath(path, options)
-  } catch (e) {
-    if (e.code !== "ENOENT") {
-      throw e;
-    }
-  }
-
-  dependOnParentDirectory(path);
-
-  return null;
-});
-
 export const optimisticReadFile = makeOptimistic("readFile", readFile);
 export const optimisticReaddir = makeOptimistic("readdir", readdir);
 export const optimisticHashOrNull = makeOptimistic("hashOrNull", (