With a smaller app on Windows, these changes reduces time spent by realpath from ~200 - 300 ms to 40ms on rebuilds, and saves ~750ms on the initial build. `realpath` works by using many lstat calls, which are slower on Windows than mac/linux.

When creating the list of files for each architecture, Meteor uses a symlink loop checker which would call `realpath` for each folder. It will now call `lstat` once for each folder and then call `realpath` for folders that are symlinks.

Meteor would also call realpath for each file in `node_modules/.bin` to check if it is a symlink. Since npm copies files into .bin instead of using symlinks on Windows, Meteor will now use lstat first, and then call `realpath` if it is a symlink. This does slightly increase the work done on linux/osx, but the additional lstat call is already cached from earlier in the build process.

The remaining use during rebuilds is calling `realpath` for each node_modules folder that is copied into the built app. I left this as is since it isn't clear why this is done and only adds ~20ms in the apps I profiled.